[ws] Fix K64F websocket build break (#1127)
Fixes #1098

+++ src/zjs_web_sockets.c
@@ -167,7 +167,7 @@ static void generate_key(char *key, uint32_t len, char *output, uint32_t olen)
     char sha_out[20];
     // compute sha1 hash of concatenated key + magic
     mbedtls_sha1(concat, strnlen(concat, 64), sha_out);
-    uint32_t out_len;
+    size_t out_len;
     // base64 encode the sha1 hash
     mbedtls_base64_encode(output, olen, &out_len, sha_out, 20);
     zjs_free(concat);
